Implement pruning techniques and UCI Clear Hash option
- Add reverse futility pruning, razor pruning, and futility pruning in `search` for low depths
- Optimize quiet move checks and LMR array index bounds safety
- Add `Clear_Hash` button handling to UCI option parsing
- Add `uci.sh` startup script for Unix environments

+++ nnue/__init__.py
@@ -18,5 +18,5 @@ _nnue.nnue_init(
 _nnue.nnue_evaluate_fen.argtypes = [ctypes.c_char_p]
 _nnue.nnue_evaluate_fen.restype = ctypes.c_int
 
-def nnue_evaluate_fen(fen: str) -> int:
+def nnue_evaluate_fen(fen: str) -> float:
     return _nnue.nnue_evaluate_fen(fen.encode("utf-8"))
